cgi script: debug fn prints array + env if we ask it
only prints one array (%) by one array, the doc that i've read seems to
point that is a source of bug and unexpected behaviors to pass more
than one associative array to a function. The function is able to print
any associative array that you pass to + the env if you set the first
arg to 1 cgi_debug(1, %_GET), will print the content of the %_GET (try
to keep the same spirit as php with the query datas) and the %ENV.

Also reset the input file balise because we need to select a filename
for being able to delete it. I've checked, it seems that it has no
incidence with the request DELETE made with the javascript snippet
for translating the request in a DELETE one.

Adding an alt attribute for checking behavior with and being more W3C
compliant