debug-upload:
Add a check line on std::cerr for looking into the buffer after a recv
inside the Listener
for confirming or not if my feelings were right about the request : I do
not have the entire body into the CGI and this is why I think upload failed

It prints the buffer on std::cerr on the server part (nothing appears in
the browser) in yellow.
It seems I'm right I do not receive the entire buffer before it is
passed to the execution, that's why I'm blocked, there is so much
paramters that could interfer this is why i do not find what is wrong

I confirm I do not receive all the body part because when you look inside
the server part you can see that I receive the Response 200 from the cgi
(white background) script which is mixed with the remaining buffer (in yellow)
I do not know if it comes from multithreading (as the read is not mutex
it may interfer but not necessary), or if it comes from the options
passed to kqueue (I've noticed that it has changed, so maybe) or if it
comes from the max_size read. Need mates' help on this