stylesheet + debug upload:
Added a snowflake animation in css, on the upload page, got struggle to
fit the animation on the size 612x612 (as the pic are) so do not add it
on the gallery part but on only one pic redim to 405px height. Maybe it
comes from the tuto I take the anim from. Do not want to spend many time
on this so this why I've done it only on one pic

For the debug upload part, it is more "delicate" because I think there
are more than one pb that fails the upload.

One of them is that we use std::string for keeping the buffer (get back
from recv) that does not keep the \0 ascii, but a png image can be
compounded with \0 ascii. So I check what we get back from the recv
buffer -> here it seems all ok apart that we get back the buffer inside
an array of char, an array of unsigned char would fit better, as for ex
a png image begins with the ascii sequence 137 80 78 71 13 10 26 10, and
after it could have \0 ascii parts. A jpeg image always begins with the
sequence of 255 216. That's why I think unsigned char would fit better.

A consequence of this is that the buffer is not faithfully reproduced
(for the png upload) as it is transferred into a std::string that would
cut the string when a \0 is reached (the buffer keeps all the \0 needed), and the cgi struggle to decode
correctly the datas (indeed it is impossible as the datas are already
corrupted before they reach the cgi part).

Also I noticed a kind of overflow with some pic, do not appear
everytime: the buffer is sent to the cgi, the cgi processes what it has
to be, but when the cgi response is sent to the server, we could see the
buffer is still reading datas that should be read before reaching the
cgi.

No jpeg has been uploaded, I do not know why but it never reaches the
cgi, do not know if it comes from the request parsing or recv is never
finished. request parsing because it "prepares the answer" only if
"request is complete". That is why I wonder.
I wonder if one solution may be to recv all the datas when it encounters
a pic upload (quite easy to define it, I've almost done it with my debug part)
or if the "plaintext" of the request shouldn't be more efficient as an
istream (for keeping the \0 ascii part). As it should impact the
structure of the server, I won't do that without mates help and approval