timeout: reintegrated previous work
previous work coming from v2-sockets-KO branch

trying to make it worked because have read that was not working into
commit, EV_ONESHOT make the timer to be non periodic.
Have setted EVFILT_TIMER into the first EV_SET (not reused the EV_SET
inside the EVFILT_READ) and create the 504 Gateway_Timeout error page +
error class and returning it if the event EVFILT_TIMER is encountered

have updated the Listener::prepare_answer for passing it the event.data
form the EVFILT_READ event -> this the pending datas size remaining,
have to upgrade the testing, that was for being sure that works

I've also read into the man page of kqueue, that if kevent returns 0 it
means that the time limit has expired -> will thorw an exc or returning
a Gateway_Timeout?