cgi manager: update exit status code error boundary
patching the code error if it misses the boundary part in the request
ugly patching but hurry patching. I can't consider all the exit status
different to 0 outputing from perl as the same script because
some errors still stay internal server error
(as script errors are - forgetting a ; these kind of things...).
The three I take out, I'm sure that are not internal server
so throwing an exception, if we encounter this, that answers with a 403
response. Others throwing a 500 reponse.